write tests for list-library
Using assertions and lemmata, tests for Lem's list library were written.
These tests revealed a mismatch between the semantics of "partition" in
OCaml and HOL 4. In HOL 4 the lists are reversed, in OCaml, they are not.
This bug, as well as some correct, but non-standard translations to Isabelle,
were fixed.