overwrite only modified files
Sometimes it is annoying that Lem overwrites existing files with their
original content. When you are working on "B.lem" which uses "A.lem",
all the files produced by "A.lem" are recreated whenever you modify
"B.lem" and run it through Lem. If "A.lem" produces lengthy OCaml
output or much OCaml code depends on it, it is annoying that this code
needs to be recompiled. Similarly, (and even more time consuming) may
be running the theorem prover output of "A" through the provers again.

This checkin contains a very simple, but not efficient or elegant
solution to this problem. Instead of writing to "A.output" directly,
Lem first writes to some temporary file "temp". Then "A.output" is
compared with "temp". If they differ, "A.output" is replaced with
"temp", otherwise the old file is kept.