cleanup representation of targets
The type of typed-ast targets was represented as "Target.target option". None represented the identity target, which is
not representable explicitly in the untyped ast. This commit renames "Target.target" to
"Target.non_ident_target" and introduces a new variant type "Target.target" to replace the option-type.
Moreover, the use of "Target.target" is cleaned-up in many places.