fix name resolution of Id_none
Ids carry either an identifier via "Id_some i" (usually as the user typed it)
or have a "Id_none" entry (if generated by some macro).
"Id_none" needs resolving into an suitable identifier. This identifier depends on the local
context. Unluckily the local context was not available. This commit adds a local context to
definitions. This context is then be used by the backends to figure out suitable identifiers.