cleanup of the type Typed_ast.env_tag
env_tag stored the type of constant as well as for which targets a constant was defined.
By moving the targets into an explicit new field "const_targets" of the record "const_descr",
a lot of code could be simplified. As part of this change, the exception "Ident.No_type",
which was used to report general type errors, was removed from
the identier module and replaced by one in "Reporting_basic".