improved pretty printing
The decision whether the whitespace should be rearranged during pretty printing is propagated via locations (type Ast.l).
If a "Trans" location was used, the whitespace was rearranged. The idea is that Trans indicated that some macro did some serious
change and therefore the original whitespace is not sensible any more. However, this turned out to be too crude. As a quick (not elegant) workaround,
I added a flag to the Trans location, stating, whether reformating should be performed.