Pass context information to the macro system for finer control of when macros fire.
Large change to fix an annoying issue in the Coq backend: Coq can handle
quantifications when they are in Prop.  When they are in bool, they need
compiling away with a macro.  When extracting a lemma statement, we wish
to remain in Prop, so quantifiers should not be compiled away.  The
macro system currently does not support passing any context information
to the macros, so they can decide whether to apply themselves or not.
This is now changed, with both pattern and expression macros being
passed an extra flag telling them whether they are in the context of a
lemma statement or not.  If this proves useful for other situations,
then it can be expanded later.