more work on the new library, warning: old lib deactivated now
Warning: This checking deactivates the old libs. This means that
no examples work currently! This should change soon :-)

There are a lot of minor changes with typechecking declarations and
pushing them to the constant description. The new-libraries got fixed
a bit to avoid additional type-checking errors that are now detectable.
The dependency resolution became more mature and now searches for
files in the library directory and the directory of the parent theory.
There are also more warnings and better error messages.