introduced indirection for modules
similar to constants and types, now the environment provides in local-environments
a map from module names to the full path of the module (m_env). The global-environment
e_env maps these paths to module descriptions. This change is in preparation of
renaming modules and changing the module structure depending on the backend.