implemented a tiny bit of search for type class instantiations
now we can have a default instantiation that is overriden later. This is
used to use "unsafe_structural_equality" as the default instantiation for
the class "Eq". However, this is only a temporary quick and dirty solution.
In the medium run, type definitions for record and variant types should automatically
generate more sophisticated default instantiations of various type classes like
"Eq" or "Ord".