Added a string library for Lem
Added a rudimentary string library for Lem, including abstract character
and string types.  Basic operations on characters (e.g. to and from
natural number codes) and lists (e.g. to and from character lists) and
supported.

Also added a pair of type classes for converting values of a given type
to and from strings in Show and Read, modelled on the Haskell
equivalents.  Instances for builtin in types are supplied.

ASSUMPTION: characters in the library are ASCII characters.