Coq backend now outputs extra files containing lemmas, assertions, etc.*
* Coq backend now outputs extra files containing lemmas, assertions and
related concepts.  If processing file $FOO, backend produces $FOO.v
containing main definitions, instances, class declarations and so on,
and $FOO_extra.v containing lemmas, assertions and so on that should not
be included in a standard file and are merely used for testing/sanity
checking definitions.  This brings the Coq backend into line with the
other theorem prover backends that already do this.

* Some other changes to the Coq backend to get the library files
correctly processed.