Fix indreln output for HOL
Added a new reln_show_types parameter to the backends so that the
relation names and types don't have to be printed out before the
clauses.  The parameter is set to false for HOL and true for all of the
other back ends.  This is correct for HOL, but I don't know about any of
the others.