implement target-specific class methods
Class methods can now be defined for certain targets only.
Typechecking ensures that this restriction is enforced when using
the typeclass. If all methods are not defined for a certain backend,
this class can be dropped for this backend.

This allows the generation of better HOL and Isabelle code. Set implementations
of OCaml and Coq need an order on the elements. HOL and Isabelle don't need
this order. Before, the HOL and Isabelle code generated for functions using
polymorphic sets, contained additional class-type dictionary arguments, which
were never used. This change eliminates these extra arguments, leading to much more
idiomatic code.

As part of this change, the Ord class is no longer used for sets and maps.
Instead special type-classes "SetType" and "MayKeyType" are introduced.