work on HOL 4 libs, now a heap is used
some steps towards getting the HOL 4 libs tested on a more regular basis.
Now a heap called "lemheap" is generated. For a template how to use this
hol heap see "library/Holmakefile". This heap is used for automatically
running the HOL-auxiliary files of the libs through HOL. However, eval
still fails often and we need to provide computeLib with some rewrite
rules for these executable functions. This especially affects maps.