improved type-checking
the typechecker now detects type-variables on the right hand side of a defintion that do not appear in the type of the resulting function.
For example

let b : bool = (exists (x : 'a) y. not (x = y))

checks, whether the type of x and y contains at least two elements. However, this type 'a does not
appear in the signature of b. However, b's value depends on this hidden argument. Therefore, this definitions and similar ones cause
trouble and are now prohibited.