generate alldoc nicer
modification of process_file and main to get the latex output in one file done nicer. Now it is possible to generate
the latex-output in just one file with the option "-tex_all filename". In contrast to before, this does not generate
single tex-files for all input modules any more. The old "-tex" option can still be used to generate output for single files.