support polymorphic number pattern
number patterns like "0", "5", but also  "x + 10" worked only for type nat before. Now they are polymorphic.
However, the types for which they work are still hard-coded in patterns.ml. Currently they work for nat and natural.
Getting them working for integers (with negative numbers present) does not make sense, since a pattern "x + ???" would match
all integers then and is better written as "x -> let x' = x - ??? ...". So, for integers we would only have the literal case split,
which probably does not justify the extra work needed to get it working. If we define more bounded natural number types in the future,
it is easy to extend the current scheme to them.