Fixed two bugs in Coq backend: mismatch between generated fixity and Coq defaults, re-export of imported files
Fixed two bugs in the Coq backend:
  * Names marked infix in Lem were previously generated with a default
fixity level and associativity.  Some names have pre-defined fixity and
associativities in Coq, which caused Coq to produce an error if this
predefined setting differed from the default being generated by Lem.  We
now have a simple (hard-coded, for the time being) mapping within Lem of
the default fixities and associativities of some common Coq names which
should be used instead of the default.  As a result, library file
num.lem now produces Coq code which no longer needs minor hand editing.

* Re-exported imported modules.  Lem assumes importation is transitive.
Coq doesn't.  As a result, imported modules need to be re-exported to
fix this mismatch.