Remove "ordering" and "set" from library/coq_constants, and add "value".
With "ordering" and "set" listed in coq_constants, the references to these types in library/*.lem were getting re-written incorrectly.

Without "value" listed in coq_constants, a variable named "value" used inside a pattern was incorrectly interpreted by Coq as being a reference to Coq.Init.Specif.value (which is an alias for the Some value constructor).