fix bugs in processing submodules
When processing submodules, target bindings were not pruned. This means
that definitions intended for other backends were still present in
submodules and caused trouble. For example, it was not possible to write
down lemmata in submodules, because the backends that don't support
lemmata otherwise complained about e.g. the quantification. Fixing this
bug revealed that checking for duplicates was buggy as well and often
throw TODO-exceptions. This is fixed now as well.