added command line argument -hol-remove-matches
There is a defintion macro that tries to turn top-level matches into
function definitions with multiple clauses. This macro is by default
activated for Isabelle and would also be very useful for HOL4. Some time
ago it was decided not to use it though for backwards-compatibility
reasons (also arguably the resulting HOL4 code is more idiomatic and
easier usuable). I now added a command line argument "-hol-remove-
matches" that enables this definition macro for HOL4. It is still
disabled by default. However, I changed the makefile to use it by
default for building the HOL libs. This should not cause trouble with
backwards compatibility, since most functions are mapped to existing HOL
definitions and the few remaining ones were not used by old versions of
Lem.