don't output full path to isabelle-lib by default
don't print the full path of the isabelle-lib for import statements in
Isabelle any more. The old behaviour can be reactivated with the option
"-add_full_isa_lib_path", though. Removing the absolute path leads to
more portable output. One needs to start Isabelle with argument "-d
path_to_isabelle_lib" though, and this dir to the local Isabelle config
or copy the lib files to the same dir as the output.