Add some list and string library functions
For lists, add a function all_distinct with a Lem implementation and a
HOL representation. It's in List_extra right now, but it should probably
move to List, once it has the appropriate target representations.

For strings, added the following to String_extra:

ord
chr
Show typeclass with instances for nat and natural
nth
stringConcat

All but ord and chr have Lem implementations and HOL representations,
some have OCaml representations too. ord and chr have representations
for HOL, Coq, Isabelle and OCaml, although the Coq and Isabelle are just
guesses based on some Googling.

ord, Show, and stringConcat can probably move to String as with
all_distinct. nth and chr are partial, and so need to stay put.