Implemented new flag -prover_remove_failwith for removing fail and failwith in match statements
At Peter and Shaked's request:

* Implemented macro to remove fail and failwith from Assert_extra.lem
when they appear in the body of a case statement.  Makes the case
statement partial, therefore pattern match compilation machinery is
invoked to replace the missing cases with default values or similar
tricks for backends.

* Added new command line flag -prover_remove_failwith in order to
activate this functionality, due to the large slowdown this entails when
using Lem on large codebases.

* Currently working with Coq and HOL.  Currently interacting badly with
pattern compilation in the Isabelle backend and is under investigation
as to how to fix this.