partly undo 26220e8
In commit 26220e8, I changed the translation of (fun _ -> ...) from
(\x. case x of _ => ...) to (\_. ...) for HOL4. This is unluckily not
valid input syntax for HOL4, if occuring in certain contexts, e.g.
inside a record definition. So, this commit undoes this change.

The resulting HOL4 code is still ugly. I played around with doing more
clever things in Lem. However, it quickly becomes complicated.
Therefore,  in my opinion the best solution is changing the Lem input
from "fun _ -> ..." to "fun _x -> ...".