Resolved issue #139
* Resolved issue 139 by making Lem much more aggressive in its
renamings.  Side effect: many more pattern variables and function
arguments are now renamed, and a global counter is maintained for
generating fresh names, rather than the complicated and partially broken
system that was in place before.

This seems to be the only workaround to issue #139 without significantly
rearchitecturing the design of Lem's renaming mechanism in order to spot
that the following example, due to Basille, is problematic:

let foo (x : nat) (X : nat) =
    let (x0 : nat) = x in
      (x, X, x0)

This resolves issue #139.