Resolve issue #152
* Resolve issue #152, a problem with the renaming mechanism taking
exponential time in the number of bound names in the term.  Lem now
caches bound names as a term is built (and does an ugly double
application of the get_renames function to figure out which bound
variables are going to be renamed once the lazy substitution is applied,
to avoid clashes between top-level pattern variables and bound variables
in an expression).  Basile's blowup.lem file is now processed
instantaneously on my machine.

This resolve issue #152.  Hopefully.