Add compilation of pattern-matching as indreln
A new [compile_relation_exp] functions is created in patterns.ml and
exposed in patterns.mli .

Contrary to other pattern compilation functions, it preserves the
semantic of a set of patterns as an inductive relation and not as a
pattern-match: overlapping patterns are kept and preserved, order do not
matter, etc.

This code may be better placed in another file as it is strongly linked
to convert_relation's architecture, but it needs to access the internals
of the pattern-compilation mechanism.

In essence, the idea is that given a set of pattern-matching
representing an inductive relation (w/ a couple of restrictions,
essentially that there are no free variables in the body of the match),
we want to convert to a real pattern-matching in a more efficient way
than evaluating all the possible matches in parallel.

To do so, we merge as much patterns as possible (e.g. [A (B x)] and
[A (C y z)] would be converted into [A tmp => match tmp with ...]
instead of branching), using two user-provided functions to perform the
tasks of branching (when a pattern is not enough to discriminate between
two inductive rules) and failing (when some pattern is not covered by
any inductive rules).

It is not used yet, but convert_relation will be updated to use it.