Properly linearize patterns in indreln->patterns
Adds a linearization pass to translate a single patterns binding
multiple times the same variable when translating from inductive
relations to patterns.

In order to do so, an internal representation of patterns for the
translation is introduced, which should help further development
on pattern simplification.

The old linearization code is still used to ensure that nested
bindings are still correctly instanciated (this modification
handles e.g. [twice (Node x x) x] with mode i -> o, but does not
handle it with mode i -> i, for which the more global old code
is still needed.