Reverse concatenation of whitespace
This seems to have been the wrong way around.  For example, in the following
Lem function, each let is followed by a newline, and some spaces for
indentation.  In the Isabelle output, Lem would print the spaces first, and
then the newline, so that the indentation is list.

val foo : integer -> integer
let foo n =
  let (bar : integer) =
    let (baz : integer) =
      if n > 0 then 47 else 11
    in baz - 42
  in bar + 42

With this patch, the whitespace is preserved.