Add some support for Isabelle's session-qualified imports
The -lib command line option for adding directories to the library path can now
be used to optionally associate a session name with a directory.  For example,

  lem -lib MyLib=path/to/mylib ...

will use the MyLib session name to qualify names of theories imported from
path/to/mylib when outputting to Isabelle.  When no session name is found for a
theory import, it is output unqualified.

Path-based instead of session-qualified imports can be enabled using the
-isa_path_imports option (replacing -isa_add_full_library_path).  Lem then adds
(full) paths to theory imports, except for theories that are output into the
same directory as the importing theory.